v 1.3.7 Refactor branch
[Seperation of comment component in expectation of forums. Further work on interfaces and tsx adoption.]
